[Refactor] migration to fastmath.py, some code simplification, renamed apis in fastmath, et al.
